lace.builtin.define: Clarify unknown control type error
There was much confusion when trying to work out why
define foo [bar] did not work.

The fact that [bar] expands to `define $RANDOM bar` and the [bar] token is
replaced by $RANDOM, which isn't a control type and is not defined until
runtime, so can't work.

Pragmatically this only prevents you using define to create an alias for
another command, which is of debatable utility.