fix bug where '[' would be lost
If the accumulator was non-empty during the lexing of a lexeme
which contained '[' such as ![foo] the ! being in the accumulator
would prevent the [foo] being processed as a quoted sublexer. Since
[] quoting is only for sublexing, instead we add the accumulator
to the lexeme so that it can be used later in the compiler to ensure
that ![foo] correctly turns into !__autodefN etc.